+++ assets/javascripts/kitten/components/interaction/upload/document-manager/stories.js
@@ -149,7 +149,7 @@ StatusError.args = {
   displayTitle: 'Extrait KBIS refusé',
   displayContent: (
     <>
-      <Text weight="light" size="tiny" tag="p" className="k-u-margin-none">
+      <Text weight="light" size="small" tag="p" className="k-u-margin-none">
         Mauvaise adresse :
       </Text>
       <BulletList